devtools/server/tests/chrome/test_framerate_02.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1007200 - Create a framerate actor
-->
<head>
  <meta charset="utf-8">
  <title>Framerate actor test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="framerate-helpers.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script>
"use strict";

window.onload = async function() {
  const target = await getTargetForSelectedTab();
  const front = await target.getFront("framerate");

  const rawData = await front.stopRecording();
  ok(rawData, "There should be a recording available.");
  is(rawData.length, 0, "...but it should be empty.");

  const timeline = plotFPS(rawData);
  is(timeline.length, 2,
    "There should be one measurement plotted, with two entries.");

  info("The framerate should be assumed to be 0 if the recording is empty.");

  is(timeline[0].delta, 0,
    "The first time delta should be 0.");
  is(timeline[0].value, 0,
    "The first framerate value should be 0.");

  is(timeline[1].delta, 100,
    "The last time delta should be 100 (the default interval value).");
  is(timeline[1].value, 0,
    "The last framerate value should be 0.");

  await target.destroy();
  SimpleTest.finish();
};
</script>
</pre>
</body>
</html>
